Added irrevocability to NOrec
 - norec now attempts to become irrevocable (by falling back to tml-like behaviour) when
 doing memory frees.
 - this simplifies a lot of things and makes reasoning about its correctness
 much easier.
    - also it now works at high contention workloads with a lot of threads
    - optimised norec when beginning (takes glb if its even otherwise rolls it back by 1)
    - not sure exactly why this works...(will have to find out)

 - i changed the benchmarks to be based on number of seconds rather than
 rounds
    - this way i can calculate retries/sec and commits/sec as a data point

 - also i refactored common code from the pmdk stms

 - also hashmap resizing works

 - added compiler memory reorder fences when accessing the glb
    - made glb and loc 64bit unsigned
    - glb shouldn't have been _Atomic either as assigning to it did synchronisation

+++ volatile_stms/algs/common/tx.c
@@ -21,8 +21,8 @@ struct tx {
 	int num_commits;
 };
 
-static _Atomic int total_retries = 0;
-static _Atomic int total_commits = 0;
+static _Atomic uint64_t total_retries = 0;
+static _Atomic uint64_t total_commits = 0;
 
 static struct tx *get_tx(void)
 {
@@ -289,7 +289,6 @@ int tx_end(void (*end_cb)(void))
 		tx->stage = TX_STAGE_WORK;
 		tx->level--;
 		if (tx->last_errnum)
-		// if ((tx->last_errnum > 0) || (tx->retry && tx->last_errnum == -1))
 			tx_abort(tx->last_errnum);
 	}
 
@@ -317,7 +316,6 @@ void tx_reclaim_frees(void)
 		entry = &tx->free_list->arr[i];
 
 		DEBUGPRINT("[%d] freeing...", tx->tid);
-		// printf("[%d] freeing...\n",tx->tid);
 		if (entry->addr != NULL)
 			free(entry->addr);
 	}
